# docs/api-contract.openapi.yaml
openapi: 3.1.0
info:
  title: 인테리어 견적 관리 API
  version: 0.2.0
  description: |
    프론트와 백엔드의 단일 진실 공급원. 어떤 코드보다 먼저 수정한다.

    ## 응답 규격
    - **성공**: `{ ok: true, data: T }`
    - **실패**: `{ ok: false, error: string, fieldErrors?: Record<string, string> }`

    ## 인증
    세션 쿠키 기반 (`sid`). `/api/health`, `/api/auth/signup`, `/api/auth/login` 외
    모든 라우트는 로그인 필요. 401 응답 시 클라이언트는 `/login` 으로 리다이렉트.

    ## 변경 절차
    1. 이 문서 먼저 수정
    2. 백엔드 라우트 + repos 구현
    3. 프론트 `api.js` 함수 + 모듈 호출
    4. 양쪽 typecheck 통과 확인

    ## 도메인 노트
    - **견적 채택**: `quotes.status` 와 `line_items.selected_quote_id` 를 의도적으로 분리.
      `POST /api/quotes/{id}/select` 는 line_items.selected_quote_id 만 갱신, status 는 사용자가 수동 관리.
    - **첨부**: 현재 base64 data URL 로 직렬화해 JSON 으로 전송, 5MB 제한.
      추후 `@fastify/multipart` + S3 로 이전 시 별도 라우트 분리.
    - **순서 변경**: phases / spaces / line-items / review-materials 는 별도 reorder 엔드포인트.
      attachments 도 동일.
  contact:
    name: Project repository
    url: https://github.com/your-org/interior-v2
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3001
    description: 로컬 개발 서버
  - url: https://interior.example.com
    description: 운영 (예시)
...
security:
  - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 세션 쿠키. signup/login 성공 시 발급. health/signup/login 외 모든 라우트는 이 쿠키 필요.
  schemas:
    ApiSuccess:
      type: object
      required: [ok, data]
      properties:
        ok: { type: boolean, enum: [true] }
        data: {}
    ApiError:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, enum: [false] }
        error: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
    DeleteSuccess:
      description: 삭제 성공 응답. data.id 는 삭제된 리소스 id.
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [id]
              properties:
                id: { type: string }
...
    User:
      type: object
      required: [id, email, name]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
...
    Attachment:
      type: object
      required: [id, projectId, ownerType, ownerId, kind, filename, blobUrl, sortOrder, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        ownerType:
          type: string
          enum: [project, vendor, phase, space, lineItem, quote, contract, payment, changeOrder, asTicket]
        ownerId: { type: string }
        kind: { type: string, enum: [photo, pdf, drawing, document, other] }
        photoKind:
          type: string
          enum: [before, during, after, reference, defect, floorplan, naver_floorplan, fixture]
          nullable: true
        filename: { type: string }
        contentType: { type: string, nullable: true }
        byteSize: { type: integer, nullable: true }
        blobUrl: { type: string }
        blobPathname: { type: string, nullable: true }
        caption: { type: string, nullable: true }
        takenAt: { type: string, format: date, nullable: true }
        linkUrl: { type: string, nullable: true }
        uploadedById: { type: string, nullable: true }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: 로그인 필요
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 리소스 없음 또는 권한 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: 입력 검증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
paths:
...
  /api/auth/signup:
    post:
      tags: [auth]
      security: []
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string, minLength: 1 }
      responses:
        "200":
          description: 가입 후 자동 로그인
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /api/auth/login:
    post:
      tags: [auth]
      security: []
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
            examples:
              normal:
                summary: 정상 로그인
                value: { email: "user@example.com", password: "secret1234" }
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401":
          description: 로그인 실패 (이메일 또는 비밀번호 불일치)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/auth/logout:
    post:
      tags: [auth]
      summary: 로그아웃
      responses:
        "200":
          description: 로그아웃됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean, enum: [true] }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/me:
    get:
      tags: [auth]
      summary: 현재 로그인 사용자
      responses:
        "200":
          description: 현재 사용자
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/projects/{projectId}/attachments:
    parameters:
      - in: path
        name: projectId
        required: true
        schema: { type: string }
    get:
      tags: [attachments]
      summary: 프로젝트 첨부 목록
      responses:
        "200":
          description: 첨부 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Attachment" }
    post:
      tags: [attachments]
      summary: 첨부 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ownerType, ownerId, filename, blobUrl]
              properties:
                ownerType:
                  type: string
                  enum: [project, vendor, phase, space, lineItem, quote, contract, payment, changeOrder, asTicket]
                ownerId: { type: string }
                kind: { type: string, enum: [photo, pdf, drawing, document, other] }
                photoKind:
                  type: string
                  enum: [before, during, after, reference, defect, floorplan, naver_floorplan, fixture]
                  nullable: true
                filename: { type: string }
                contentType: { type: string, nullable: true }
                byteSize: { type: integer, nullable: true }
                blobUrl: { type: string }
                blobPathname: { type: string, nullable: true }
                caption: { type: string, nullable: true }
                takenAt: { type: string, format: date, nullable: true }
                linkUrl: { type: string, nullable: true }
      responses:
        "200":
          description: 생성된 첨부
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Attachment" }
  /api/projects/{projectId}/attachments/reorder:
    parameters:
      - in: path
        name: projectId
        required: true
        schema: { type: string }
    patch:
      tags: [attachments]
      summary: 첨부(사진) 순서 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: 재정렬된 첨부 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Attachment" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/attachments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [attachments]
      summary: 첨부 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind: { type: string, enum: [photo, pdf, drawing, document, other] }
                photoKind:
                  type: string
                  enum: [before, during, after, reference, defect, floorplan, naver_floorplan, fixture]
                  nullable: true
                filename: { type: string }
                contentType: { type: string, nullable: true }
                byteSize: { type: integer, nullable: true }
                blobUrl: { type: string }
                blobPathname: { type: string, nullable: true }
                caption: { type: string, nullable: true }
                takenAt: { type: string, format: date, nullable: true }
                linkUrl: { type: string, nullable: true }
      responses:
        "200":
          description: 수정된 첨부
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Attachment" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [attachments]
      summary: 첨부 삭제
      responses:
        "200":
          description: 삭제됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteSuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...